modularização e tratativas de erros
